+++ andes/system.py
@@ -143,7 +143,7 @@ class System:
                                      ('dime_address', 'ipc:///tmp/dime2'),
                                      ('numba', 0),
                                      ('numba_parallel', 0),
-                                     ('numba_nopython', 0),
+                                     ('numba_nopython', 1),
                                      ('yapf_pycode', 0),
                                      ('save_stats', 0),
                                      ('np_divide', 'warn'),
@@ -824,7 +824,7 @@ class System:
 
         self.s_update_post(models)
 
-        # store the inverse of time constants
+        # store time constants associated with differential equations
         self._store_tf(models)
 
     def store_adder_setter(self, models):
@@ -1255,6 +1255,7 @@ class System:
                         to + fr + fr + to,
                         (n, n),
                         'd')
+        temp = sparse(temp)  # need to drop allocated zero values
 
         cons = temp[0, :]
         nelm = len(cons.J)